+++ packages/google_maps_flutter/google_maps_flutter_ios/CHANGELOG.md
@@ -1,3 +1,15 @@
+## NEXT
+
+* Updates minimum supported SDK version to Flutter 3.19/Dart 3.3.
+
+## 2.12.0
+
+* Adds support for marker clustering.
+
+## 2.11.0
+
+* Adds support for heatmap layers.
+
 ## 2.10.0
 
 * Converts Obj-C->Dart calls to Pigeon.